+++ sdk/identity/azure_identity/src/imds_managed_identity_credential.rs
@@ -136,28 +136,28 @@ impl ImdsManagedIdentityCredential {
             )
             .await?;
 
-        if !rsp.status().is_success() {
-            match rsp.status() {
+        let status = rsp.status();
+        if !status.is_success() {
+            let message = match status {
                 StatusCode::BadRequest => {
-                    return Err(Error::with_message(
-                        ErrorKind::Credential,
-                        "the requested identity has not been assigned to this resource",
-                    ))
+                    "The requested identity has not been assigned to this resource".to_string()
                 }
                 StatusCode::BadGateway | StatusCode::GatewayTimeout => {
-                    return Err(Error::with_message(
-                        ErrorKind::Credential,
-                        "the request failed due to a gateway error",
-                    ))
+                    "The request failed due to a gateway error".to_string()
                 }
                 _ => {
                     let body = String::from_utf8_lossy(rsp.body());
-                    return Err(Error::with_message(
-                        ErrorKind::Credential,
-                        format!("the request failed: {body}"),
-                    ));
+                    format!("The request failed: {body}")
                 }
-            }
+            };
+            return Err(Error::new(
+                ErrorKind::HttpResponse {
+                    error_code: None,
+                    raw_response: Some(Box::new(rsp)),
+                    status,
+                },
+                message,
+            ));
         }
 
         let token_response: MsiTokenResponse = from_json(rsp.into_body())?;