+++ sdk/identity/azure_identity/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- A `get_token()` error motivated by an HTTP response carries that response. See the [troubleshooting guide](http://aka.ms/azsdk/rust/identity/troubleshoot#find-relevant-information-in-errors) for example code showing how to access the response.
+
 ### Breaking Changes
 
 - `ClientCertificateCredential::new()`: